+++ app/assets/javascripts/govuk_publishing_components/analytics-ga4/ga4-scroll-tracker.js
@@ -115,7 +115,7 @@ window.GOVUK.Modules = window.GOVUK.Modules || {};
         data.type = node.eventData.type
         // following will be undefined if tracking percentages
         data.text = node.eventData.text
-        data.index = window.GOVUK.analyticsGa4.core.ensureIndexesArePopulated(node.eventData.index || {})
+        data.index = node.eventData.index
         // following will be undefined if tracking headings
         data.percent_scrolled = node.eventData.percent_scrolled