+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ## Unreleased
 
 * Ensure search forms have their text sent to GA4 in lowercase ([PR #3504](https://github.com/alphagov/govuk_publishing_components/pull/3504))
+* Add GA4 HTML attachment tracking to attachment component ([PR #3500](https://github.com/alphagov/govuk_publishing_components/pull/3500))
 
 ## 35.11.0