+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Add GA4 HTML attachment tracking to attachment component ([PR #3500](https://github.com/alphagov/govuk_publishing_components/pull/3500))
 * Add functionality for preventing the redaction of publicly available information ([PR #3509](https://github.com/alphagov/govuk_publishing_components/pull/3509))
 * Ensure extra spaces are removed from search forms ([PR #3512](https://github.com/alphagov/govuk_publishing_components/pull/3512))
+* Add link option to tabs component ([PR #3486](https://github.com/alphagov/govuk_publishing_components/pull/3486))
 
 ## 35.11.0